updates design + removing nativewind
Nativewind in the project wasn't working correctly with flex, and I had a big problem that I spent hours trying to fix.

Maybe the problem was my configuration, maybe I should study more about this lib in the future, since the concept of className is much faster than doing it directly with style + stylesheet.

However, this inconsistency made me lose many hours of development, which for me is the worst of all, that's why I'm removing it!